streamlit/requirements.txt
streamlit==1.29.0
google-cloud-bigquery==3.13.0
plotly==5.17.0
pandas==2.1.4
db-dtypes==1.2.0